requirements.txt
aiogram==2.25
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coloredlogs==15.0.1
flatbuffers==23.5.26
frozenlist==1.4.1
humanfriendly==10.0
idna==3.6
imageio==2.34.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy_loader==0.3
llvmlite==0.42.0
magic-filter==1.0.12
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
onnxruntime==1.17.1
opencv-python-headless==4.9.0.80
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
protobuf==4.25.3
PyMatting==1.1.12
pyreadline3==3.4.1
pytz==2024.1
referencing==0.33.0
rembg==2.0.55
requests==2.31.0
rpds-py==0.18.0
scikit-image==0.22.0
scipy==1.12.0
sympy==1.12
tifffile==2024.2.12
tqdm==4.66.2
urllib3==2.2.1
yarl==1.9.4
